building three different applications, one of which dlopens the ipc library
