+++ sources/common/error_handling.cpp
@@ -1,5 +1,7 @@
 #include "error_handling.h"
-#include <cuda.h>
+#if defined(HAVE_CUDA_DRIVER) && HAVE_CUDA_DRIVER
+    #include <cuda.h>
+#endif
 #include "spdlog/spdlog.h"
 #include "ipc.pb.h"
 #include "zmq.hpp"
@@ -17,11 +19,11 @@ static const char* statusToStr(ipc::Status s) {
     }
 }
 
-int handleIpcError(ipc::Status status, const char* file, int line, const char* func, const char* msg) {
+int handleIpcError(ipc::Status status) {
     if (status != ipc::ST_SUCCESS) {
         const char* errorName = statusToStr(status);
         printf("IPC Error: [%s]\n", errorName);
-        spdlog::error("IPC Error: [{}] | File: {} | Line: {} | Function: {} | Message: {}", errorName, file, line, func, msg ? msg : "NONE");
+        spdlog::error("IPC Error: [{}]", errorName);
         return static_cast<int>(status);
     }
 
@@ -37,22 +39,24 @@ int handleZmqSendError(zmq::send_result_t status, const char* file, int line, co
     return EC_SUCCESS;
 }
 
-int handleCUDAError(CUresult status, const char* file, int line, const char* func, const char* msg) {
-    if (status != CUDA_SUCCESS) {
-        const char* errorName = "Unknown CUDA Error";
-        const char* errorString = "No additional information";
+#if defined(HAVE_CUDA_DRIVER) && HAVE_CUDA_DRIVER
+    int handleCUDAError(CUresult status, const char* file, int line, const char* func, const char* msg) {
+        if (status != CUDA_SUCCESS) {
+            const char* errorName = "Unknown CUDA Error";
+            const char* errorString = "No additional information";
 
-        cuGetErrorName(status, &errorName);
-        cuGetErrorString(status, &errorString);
+            cuGetErrorName(status, &errorName);
+            cuGetErrorString(status, &errorString);
 
-        spdlog::error("CUDA Error: [{} - {}] | Description: {} | File: {} | Line: {} | Function: {} | Additional Info: {}",
-            errorName, (int)status, errorString, file, line, func, msg ? msg : "NONE");
+            spdlog::error("CUDA Error: [{} - {}] | Description: {} | File: {} | Line: {} | Function: {} | Additional Info: {}",
+                errorName, (int)status, errorString, file, line, func, msg ? msg : "NONE");
 
-        return static_cast<int>(status);
+            return static_cast<int>(status);
+        }
+
+        return EC_SUCCESS;
     }
-
-    return EC_SUCCESS;
-}
+#endif
 
 int handleRegularError(int status, const char* file, int line, const char* func, const char* msg) {
     if (status != EC_SUCCESS) {
@@ -67,9 +71,14 @@ int errorCheck(StatusType status, const char* file, int line, const char* func, 
     if constexpr (Type == ErrorType::ZMQ_SEND) {
         return handleZmqSendError(status, file, line, func, msg);
     } else if constexpr (Type == ErrorType::IPC) {
-        return handleIpcError(status, file, line, func, msg);
+        return handleIpcError(status);
     } else if constexpr (Type == ErrorType::CUDA) {
-        return handleCUDAError(status, file, line, func, msg);
+        #if defined(HAVE_CUDA_DRIVER) && HAVE_CUDA_DRIVER
+            return handleCUDAError(status, file, line, func, msg);
+        #else
+            spdlog::error("CUDA support is not available in this build.");
+            return EC_FAILURE;
+        #endif
     } else if constexpr (Type == ErrorType::DEFAULT) {
         return handleRegularError(status, file, line, func, msg);
     } else {
@@ -79,5 +88,7 @@ int errorCheck(StatusType status, const char* file, int line, const char* func, 
 
 template int errorCheck<ErrorType::ZMQ_SEND, zmq::send_result_t>(zmq::send_result_t status, const char* file, int line, const char* func, const char* msg);
 template int errorCheck<ErrorType::IPC, ipc::Status>(ipc::Status status, const char* file, int line, const char* func, const char* msg);
-template int errorCheck<ErrorType::CUDA, CUresult>(CUresult status, const char* file, int line, const char* func, const char* msg);
+#if defined(HAVE_CUDA_DRIVER) && HAVE_CUDA_DRIVER
+    template int errorCheck<ErrorType::CUDA, CUresult>(CUresult status, const char* file, int line, const char* func, const char* msg);
+#endif
 template int errorCheck<ErrorType::DEFAULT, int>(int status, const char* file, int line, const char* func, const char* msg);
